+++ HISTORY.md
@@ -13,6 +13,9 @@ Crass Change History
 * Teensy tiny speed and memory usage improvements that you almost certainly
   won't notice.
 
+* Fixed: A semicolon following a `@charset` rule would be omitted during
+  serialization.
+
 * Fixed: A multibyte char at the beginning of an id token could trigger an
   encoding error because `StringScanner#peek` is a jerkface.